+++ reserves_project/utils/__init__.py
@@ -1,5 +1,5 @@
 """Utility helpers."""
 
-from .run_manifest import write_run_manifest
+from .run_manifest import write_run_manifest, write_latest_pointer
 
-__all__ = ["write_run_manifest"]
+__all__ = ["write_run_manifest", "write_latest_pointer"]